app/requirements.txt
pymupdf==1.25.1
python-dotenv==1.0.1
openai==1.57.2
torch
transformers